+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -1,8 +1,11 @@
 import { argentWallet } from './argentWallet/argentWallet';
+import { bitKeepWallet } from './bitKeepWallet/bitKeepWallet';
 import { bitskiWallet } from './bitskiWallet/bitskiWallet';
 import { braveWallet } from './braveWallet/braveWallet';
 import { coinbaseWallet } from './coinbaseWallet/coinbaseWallet';
+import { coreWallet } from './coreWallet/coreWallet';
 import { dawnWallet } from './dawnWallet/dawnWallet';
+import { frontierWallet } from './frontierWallet/frontierWallet';
 import { imTokenWallet } from './imTokenWallet/imTokenWallet';
 import { injectedWallet } from './injectedWallet/injectedWallet';
 import { ledgerWallet } from './ledgerWallet/ledgerWallet';
@@ -14,7 +17,10 @@ import { phantomWallet } from './phantomWallet/phantomWallet';
 import { rabbyWallet } from './rabbyWallet/rabbyWallet';
 import { rainbowWallet } from './rainbowWallet/rainbowWallet';
 import { safeWallet } from './safeWallet/safeWallet';
+import { safeheronWallet } from './safeheronWallet/safeheronWallet';
 import { tahoWallet } from './tahoWallet/tahoWallet';
+import { talismanWallet } from './talismanWallet/talismanWallet';
+import { tokenPocketWallet } from './tokenPocketWallet/tokenPocketWallet';
 import { trustWallet } from './trustWallet/trustWallet';
 import { walletConnectWallet } from './walletConnectWallet/walletConnectWallet';
 import { xdefiWallet } from './xdefiWallet/xdefiWallet';
@@ -22,10 +28,13 @@ import { zerionWallet } from './zerionWallet/zerionWallet';
 
 export {
   argentWallet,
+  bitKeepWallet,
   bitskiWallet,
   braveWallet,
   coinbaseWallet,
+  coreWallet,
   dawnWallet,
+  frontierWallet,
   imTokenWallet,
   injectedWallet,
   ledgerWallet,
@@ -36,7 +45,10 @@ export {
   rabbyWallet,
   rainbowWallet,
   safeWallet,
+  safeheronWallet,
   tahoWallet,
+  talismanWallet,
+  tokenPocketWallet,
   trustWallet,
   okxWallet,
   walletConnectWallet,