+++ packages/create-rainbowkit/generated-test-app/CHANGELOG.md
@@ -1,5 +1,25 @@
 # generated-test-app
 
+## 0.1.62
+
+### Patch Changes
+
+- Updated dependencies [f1e98e84]
+- Updated dependencies [d303a3b9]
+  - @rainbow-me/rainbowkit@1.0.7
+
+## 0.1.61
+
+### Patch Changes
+
+- Updated dependencies [dc3cd10b]
+- Updated dependencies [c251d55d]
+- Updated dependencies [d5b3bd19]
+- Updated dependencies [66e84239]
+- Updated dependencies [1b4f142e]
+- Updated dependencies [e089ab98]
+  - @rainbow-me/rainbowkit@1.0.6
+
 ## 0.1.60
 
 ### Patch Changes